refactored Root.js by conditionally rendering components based on environment to simplify code and reduce number of files from 3 to 1
